requirements.txt
hvac==0.8.2
loguru==0.2.5
kubernetes==9.0.0
PyYAML==5.1
